fix: useWebSocketSchedules 활용 코드로 변경 후 렌더링 타이밍 버그 해결 (#66)
