+++ src/pages/Planner/components/ScheduleManager.tsx
@@ -4,58 +4,32 @@ import PlanButtons from "@components/schedule/PlanButtons";
 import RouteCardList from "@components/schedule/RouteCardList";
 import ScheduleBar from "@components/schedule/ScheduleBar";
 import HorisontalLogo from "@components/assets/images/HorisontalLogo";
-import { useScheduleWS } from "../hooks/ScheduleWS";
+import { useSchedules } from "../hooks/useWebSocketSchedules";
 import { ROUTES } from "@routes/type";
 import { useTypedParams } from "@common/hooks/useTypedParams";
-// import { WebSocketClient } from "@common/services/WebSocketClient";
-
-// 전체 스케줄을 나타내는 타입 정의
-// interface Schedule {
-//   scheduleId: number;
-//   planType: string;
-//   actualDate: string;
-// }
-
-// 스케줄 아이템의 타입 정의
-export interface ScheduleItem {
-  scheduleItemId: number;
-  markerId: number;
-  name?: string;
-  address: string;
-  content?: string;
-  createdAt: string;
-  updatedAt: string;
-  itemOrder: number; // 추가된 필드
-}
 
 const ScheduleManager = () => {
-  // URL 파라미터로부터 방 ID를 받아오기
-  const { roomId } = useTypedParams<typeof ROUTES.ENTER_MODAL>();
+  const { roomId } = useTypedParams<typeof ROUTES.PLANNER>();
 
-  // WebSocket을 통한 스케줄 및 스케줄 아이템 데이터와 연결 상태 관리
-  //scheduleItems 안쓰여서 뺌
   const {
     schedules,
-    isConnected,
-    requestSchedules,
+    scheduleItems,
+    requestScheduleList,
     requestScheduleItems,
     updateScheduleItem,
     deleteScheduleItem,
-  } = useScheduleWS(roomId);
+    isConnected,
+  } = useSchedules(roomId);
 
-  // 현재 날짜 및 플랜을 관리하는 상태 변수
+  // 현재 날짜 및 플랜 관리
   const [currentDate, setCurrentDate] = useState(() => schedules[0]?.actualDate || "");
   const [currentPlan, setCurrentPlan] = useState("A");
-
-  // 현재 선택된 날짜와 플랜에 해당하는 스케줄 아이템 상태 관리
-  const [currentScheduleItems, setCurrentScheduleItems] = useState<ScheduleItem[]>([]);
-
-  const [loading, setLoading] = useState(false); // 로딩 상태 추가
+  const [loading, setLoading] = useState(false);
 
   // 모든 스케줄에서 날짜만 추출한 배열
   const dateArray = schedules.map((schedule) => schedule.actualDate);
 
-  // 주어진 날짜와 플랜에 해당하는 스케줄 ID를 찾아 반환하는 함수
+  // 주어진 날짜와 플랜에 해당하는 스케줄 ID를 찾아 반환
   const findScheduleId = useCallback(
     (actualDate: string, planType: string): number | null => {
       const schedule = schedules.find(
@@ -71,13 +45,9 @@ const ScheduleManager = () => {
   const loadScheduleItems = useCallback(
     (scheduleId: number | null) => {
       if (!scheduleId) {
-        setCurrentScheduleItems([]); // scheduleId가 없으면 빈 배열로 초기화
-
         return;
       }
-
-      setLoading(true);
-      requestScheduleItems(scheduleId); // scheduleId로 아이템을 요청
+      requestScheduleItems(scheduleId); // scheduleId로 아이템 요청
     },
     [requestScheduleItems]
   );
@@ -94,85 +64,57 @@ const ScheduleManager = () => {
   const handlePlanChange = (planType: string) => {
     setCurrentPlan(planType); // 플랜 변경
     const scheduleId = findScheduleId(currentDate, planType);
-    // 변경된 플랜에 맞는 스케줄 ID 찾기
 
     loadScheduleItems(scheduleId);
   };
 
   // 날짜 변경 시 처리하는 함수
   const handleDateChange = (actualDate: string) => {
-    setCurrentDate(actualDate); // 날짜 변경
+    setCurrentDate(actualDate);
     const scheduleId = findScheduleId(actualDate, currentPlan);
-    // 변경된 날짜에 맞는 스케줄 ID 찾기
 
     loadScheduleItems(scheduleId);
   };
 
-  // 컴포넌트가 마운트될 때 LIST_SCHEDULES 메시지를 서버로 전송하여 스케줄 목록을 요청하는 함수
   useEffect(() => {
-    if (roomId) {
-      requestSchedules(); // 페이지 진입 시 스케줄 요청
+    if (roomId && isConnected) {
+      requestScheduleList();
     }
-  }, [roomId, requestSchedules]);
-
-  useEffect(() => {
-    if (isConnected) {
-      setLoading(false); // 연결 성공 시 로딩 상태 해제
-    }
-  }, [isConnected]);
+  }, [roomId, isConnected, requestScheduleList]);
 
   const onSave = useCallback(
     (scheduleItemId: number, name: string, content: string) => {
-      if (!scheduleItemId || !name || !content) {
-        // console.warn("스케줄 아이템 수정 실패: 필수 값 부족");
+      if (!scheduleItemId || !name || !content) return;
 
-        return;
-      }
-
-      // previousItemId와 nextItemId는 null로 설정하여 전달
       const previousItemId = null;
       const nextItemId = null;
 
-      // updateScheduleItem 호출
       updateScheduleItem(scheduleItemId, name, content, previousItemId, nextItemId);
     },
     [updateScheduleItem]
   );
 
   const handleDelete = (scheduleItemId: number) => {
-    setLoading(true); // 삭제 시 로딩 상태 설정
-    deleteScheduleItem(scheduleItemId); // WebSocket을 통해 삭제 요청
-
-    // 삭제 후, 서버에서 최신 아이템 목록을 다시 요청하여 UI 업데이트
+    setLoading(true);
+    deleteScheduleItem(scheduleItemId);
     const scheduleId = findScheduleId(currentDate, currentPlan);
-
-    loadScheduleItems(scheduleId); // 삭제된 후에 다시 아이템을 로드
+    
+    loadScheduleItems(scheduleId);
   };
 
   return (
     <Container>
-      {/* 플랜 버튼: 플랜 변경 시 해당 플랜에 맞는 일정 필터링 */}
       <PlanButtons currentPlan={currentPlan} onChangePlan={handlePlanChange} />
-
       <ScheduleBox>
-        {/* 일정 바: 날짜 변경 시 해당 날짜에 맞는 일정 필터링 */}
         <ScheduleBar
           currentDate={currentDate}
           schedules={dateArray} // 날짜 배열만 전달
           onChangeDate={handleDateChange}
         />
-        {/* 일정 카드 리스트: 선택된 날짜와 플랜에 맞는 스케줄 아이템을 리스트로 표시 */}
-        <RouteCardList
-          items={currentScheduleItems}
-          onSave={onSave}
-          deleteScheduleItem={handleDelete}
-        />
-        {/* 상태로 관리된 scheduleItems 전달 */}
+        <RouteCardList items={scheduleItems} onSave={onSave} deleteScheduleItem={handleDelete} />
         <button aria-hidden={loading}></button>
       </ScheduleBox>
-
       <LogoWrapper>
-        {/* 로고 표시 */}
         <HorisontalLogo />
       </LogoWrapper>
     </Container>
